refactor server code into 1 class per api
