+++ grpc/src/main/java/com/replica/RouterImpl.java
@@ -18,69 +18,31 @@
 
 package com.replica;
 
-import com.google.common.collect.Lists;
-import com.google.common.collect.Sets;
-import com.google.protobuf.Timestamp;
-import com.google.rpc.Code;
-import com.google.rpc.Status;
-import com.graphhopper.*;
-import com.graphhopper.config.Profile;
+import com.graphhopper.GraphHopper;
 import com.graphhopper.gtfs.PtRouter;
-import com.graphhopper.gtfs.Request;
-import com.graphhopper.isochrone.algorithm.ContourBuilder;
 import com.graphhopper.isochrone.algorithm.JTSTriangulator;
-import com.graphhopper.isochrone.algorithm.ShortestPathTree;
-import com.graphhopper.isochrone.algorithm.Triangulator;
-import com.graphhopper.routing.*;
-import com.graphhopper.routing.querygraph.QueryGraph;
-import com.graphhopper.routing.util.DefaultEdgeFilter;
-import com.graphhopper.routing.util.EdgeFilter;
-import com.graphhopper.routing.util.FlagEncoder;
-import com.graphhopper.routing.util.TraversalMode;
-import com.graphhopper.routing.weighting.Weighting;
-import com.graphhopper.storage.Graph;
-import com.graphhopper.storage.index.LocationIndex;
-import com.graphhopper.storage.index.Snap;
-import com.graphhopper.util.DistanceCalcEarth;
+import com.graphhopper.routing.MatrixAPI;
 import com.graphhopper.storage.GraphHopperStorage;
-import com.graphhopper.util.PMap;
-import com.graphhopper.util.Parameters;
-import com.graphhopper.util.exceptions.PointNotFoundException;
-import com.graphhopper.util.shapes.GHPoint;
+import com.replica.api.IsochroneRouter;
+import com.replica.api.MatrixRouter;
+import com.replica.api.StreetRouter;
+import com.replica.api.TransitRouter;
 import com.timgroup.statsd.StatsDClient;
-import io.grpc.protobuf.StatusProto;
 import io.grpc.stub.StreamObserver;
-import org.locationtech.jts.geom.*;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import router.RouterOuterClass.*;
-import router.RouterOuterClass.Point;
 
-import java.time.Duration;
-import java.time.Instant;
-import java.util.*;
-import java.util.function.ToDoubleFunction;
-import java.util.stream.Collectors;
-
-import static com.graphhopper.routing.util.TraversalMode.EDGE_BASED;
-import static com.graphhopper.routing.util.TraversalMode.NODE_BASED;
-import static com.graphhopper.util.Parameters.Routing.INSTRUCTIONS;
-import static java.util.stream.Collectors.*;
+import java.util.Arrays;
+import java.util.List;
+import java.util.Map;
 
 public class RouterImpl extends router.RouterGrpc.RouterImplBase {
 
-    final Set<Integer> STREET_BASED_ROUTE_TYPES = Sets.newHashSet(0, 3, 5);
-
-    private static final Logger logger = LoggerFactory.getLogger(RouterImpl.class);
     private final GraphHopper graphHopper;
-    private final PtRouter ptRouter;
-    private final MatrixAPI matrixAPI;
-    private Map<String, String> gtfsLinkMappings;
-    private Map<String, List<String>> gtfsRouteInfo;
-    private Map<String, String> gtfsFeedIdMapping;
-    private final StatsDClient statsDClient;
-    private String regionName;
-    private Triangulator isochroneTriangulator;
+
+    private StreetRouter streetRouter;
+    private TransitRouter transitRouter;
+    private MatrixRouter matrixRouter;
+    private IsochroneRouter isochroneRouter;
 
     public RouterImpl(GraphHopper graphHopper, PtRouter ptRouter, MatrixAPI matrixAPI,
                       Map<String, String> gtfsLinkMappings,
@@ -89,196 +51,11 @@ public class RouterImpl extends router.RouterGrpc.RouterImplBase {
                       StatsDClient statsDClient,
                       String regionName) {
         this.graphHopper = graphHopper;
-        this.ptRouter = ptRouter;
-        this.matrixAPI = matrixAPI;
-        this.gtfsLinkMappings = gtfsLinkMappings;
-        this.gtfsRouteInfo = gtfsRouteInfo;
-        this.gtfsFeedIdMapping = gtfsFeedIdMapping;
-        this.statsDClient = statsDClient;
-        this.regionName = regionName;
-        this.isochroneTriangulator = new JTSTriangulator(graphHopper.getRouterConfig());
-    }
-
-    @Override
-    public void routeStreetMode(StreetRouteRequest request, StreamObserver<StreetRouteReply> responseObserver) {
-        long startTime = System.currentTimeMillis();
-
-        GHRequest ghRequest = new GHRequest(
-                request.getPointsList().stream().map(p -> new GHPoint(p.getLat(), p.getLon())).collect(Collectors.toList())
-        );
-        ghRequest.setProfile(request.getProfile());
-        ghRequest.setLocale(Locale.US);
-        ghRequest.setPathDetails(Lists.newArrayList("stable_edge_ids", "time"));
-
-        PMap hints = new PMap();
-        hints.putObject(INSTRUCTIONS, false);
-        if (request.getAlternateRouteMaxPaths() > 1) {
-            ghRequest.setAlgorithm("alternative_route");
-            hints.putObject("alternative_route.max_paths", request.getAlternateRouteMaxPaths());
-            hints.putObject("alternative_route.max_weight_factor", request.getAlternateRouteMaxWeightFactor());
-            hints.putObject("alternative_route.max_share_factor", request.getAlternateRouteMaxShareFactor());
-        }
-        ghRequest.getHints().putAll(hints);
-
-        try {
-            GHResponse ghResponse = graphHopper.route(ghRequest);
-            if (ghResponse.hasErrors()) {
-                String message = "Path could not be found between "
-                        + ghRequest.getPoints().get(0).lat + "," + ghRequest.getPoints().get(0).lon + " to "
-                        + ghRequest.getPoints().get(1).lat + "," + ghRequest.getPoints().get(1).lon;
-                // logger.warn(message);
-
-                double durationSeconds = (System.currentTimeMillis() - startTime) / 1000.0;
-                String[] tags = {"mode:" + request.getProfile(), "api:grpc", "routes_found:false"};
-                tags = applyRegionName(tags, regionName);
-                sendDatadogStats(statsDClient, tags, durationSeconds);
-
-                Status status = Status.newBuilder()
-                        .setCode(Code.NOT_FOUND.getNumber())
-                        .setMessage(message)
-                        .build();
-                responseObserver.onError(StatusProto.toStatusRuntimeException(status));
-            } else {
-                StreetRouteReply.Builder replyBuilder = StreetRouteReply.newBuilder();
-                for (ResponsePath responsePath : ghResponse.getAll()) {
-                    List<String> pathStableEdgeIds = responsePath.getPathDetails().get("stable_edge_ids").stream()
-                            .map(pathDetail -> (String) pathDetail.getValue())
-                            .collect(Collectors.toList());
-
-                    List<Long> edgeTimes = responsePath.getPathDetails().get("time").stream()
-                            .map(pathDetail -> (Long) pathDetail.getValue())
-                            .collect(Collectors.toList());
-
-                    replyBuilder.addPaths(StreetPath.newBuilder()
-                            .setDurationMillis(responsePath.getTime())
-                            .setDistanceMeters(responsePath.getDistance())
-                            .addAllStableEdgeIds(pathStableEdgeIds)
-                            .addAllEdgeDurationsMillis(edgeTimes)
-                            .setPoints(responsePath.getPoints().toLineString(false).toString())
-                    );
-                }
-
-                double durationSeconds = (System.currentTimeMillis() - startTime) / 1000.0;
-                String[] tags = {"mode:" + request.getProfile(), "api:grpc", "routes_found:true"};
-                tags = applyRegionName(tags, regionName);
-                sendDatadogStats(statsDClient, tags, durationSeconds);
-
-                responseObserver.onNext(replyBuilder.build());
-                responseObserver.onCompleted();
-            }
-        } catch (Exception e) {
-            String message = "GH internal error! Path could not be found between "
-                    + ghRequest.getPoints().get(0).lat + "," + ghRequest.getPoints().get(0).lon + " to "
-                    + ghRequest.getPoints().get(1).lat + "," + ghRequest.getPoints().get(1).lon;
-            logger.error(message, e);
-
-            double durationSeconds = (System.currentTimeMillis() - startTime) / 1000.0;
-            String[] tags = {"mode:" + request.getProfile(), "api:grpc", "routes_found:error"};
-            tags = applyRegionName(tags, regionName);
-            sendDatadogStats(statsDClient, tags, durationSeconds);
-
-            Status status = Status.newBuilder()
-                    .setCode(Code.INTERNAL.getNumber())
-                    .setMessage(message)
-                    .build();
-            responseObserver.onError(StatusProto.toStatusRuntimeException(status));
-        }
-    }
-
-    // TODO: Clean up code based on fix-it comments in PR #26
-    @Override
-    public void routeMatrix(MatrixRouteRequest request, StreamObserver<MatrixRouteReply> responseObserver) {
-        long startTime = System.currentTimeMillis();
-
-        List<GHPoint> fromPoints = request.getFromPointsList().stream()
-                .map(p -> new GHPoint(p.getLat(), p.getLon())).collect(toList());
-        List<GHPoint> toPoints = request.getToPointsList().stream()
-                .map(p -> new GHPoint(p.getLat(), p.getLon())).collect(toList());
-
-        GHMRequest ghMatrixRequest = new GHMRequest();
-        ghMatrixRequest.setFromPoints(fromPoints);
-        ghMatrixRequest.setToPoints(toPoints);
-        ghMatrixRequest.setOutArrays(new HashSet<>(request.getOutArraysList()));
-        ghMatrixRequest.setProfile(request.getMode());
-        ghMatrixRequest.setFailFast(request.getFailFast());
-
-        try {
-            GHMResponse ghMatrixResponse = matrixAPI.calc(ghMatrixRequest);
-
-            if (ghMatrixRequest.getFailFast() && ghMatrixResponse.hasInvalidPoints()) {
-                MatrixErrors matrixErrors = new MatrixErrors();
-                matrixErrors.addInvalidFromPoints(ghMatrixResponse.getInvalidFromPoints());
-                matrixErrors.addInvalidToPoints(ghMatrixResponse.getInvalidToPoints());
-                throw new MatrixCalculationException(matrixErrors);
-            }
-            int from_len = ghMatrixRequest.getFromPoints().size();
-            int to_len = ghMatrixRequest.getToPoints().size();
-            List<List<Long>> timeList = new ArrayList(from_len);
-            List<Long> timeRow;
-            List<List<Long>> distanceList = new ArrayList(from_len);
-            List<Long> distanceRow;
-            Iterator<MatrixElement> iter = ghMatrixResponse.getMatrixElementIterator();
-            MatrixErrors matrixErrors = new MatrixErrors();
-            StringBuilder debugBuilder = new StringBuilder();
-            debugBuilder.append(ghMatrixResponse.getDebugInfo());
-
-            for(int fromIndex = 0; fromIndex < from_len; ++fromIndex) {
-                timeRow = new ArrayList(to_len);
-                timeList.add(timeRow);
-                distanceRow = new ArrayList(to_len);
-                distanceList.add(distanceRow);
-
-                for(int toIndex = 0; toIndex < to_len; ++toIndex) {
-                    if (!iter.hasNext()) {
-                        throw new IllegalStateException("Internal error, matrix dimensions should be " + from_len + "x" + to_len + ", but failed to retrieve element (" + fromIndex + ", " + toIndex + ")");
-                    }
-
-                    MatrixElement element = iter.next();
-                    if (!element.isConnected()) {
-                        matrixErrors.addDisconnectedPair(element.getFromIndex(), element.getToIndex());
-                    }
-
-                    if (ghMatrixRequest.getFailFast() && matrixErrors.hasDisconnectedPairs()) {
-                        throw new MatrixCalculationException(matrixErrors);
-                    }
-
-                    long time = element.getTime();
-                    timeRow.add(time == Long.MAX_VALUE ? -1 : Math.round((double)time / 1000.0D));
-
-                    double distance = element.getDistance();
-                    distanceRow.add(distance == Double.MAX_VALUE ? -1 : Math.round(distance));
-
-                    debugBuilder.append(element.getDebugInfo());
-                }
-            }
-
-            List<MatrixRow> timeRows = timeList.stream()
-                    .map(row -> MatrixRow.newBuilder().addAllValues(row).build()).collect(toList());
-            List<MatrixRow> distanceRows = distanceList.stream()
-                    .map(row -> MatrixRow.newBuilder().addAllValues(row).build()).collect(toList());
-
-            double durationSeconds = (System.currentTimeMillis() - startTime) / 1000.0;
-            String[] tags = {"mode:" + request.getMode() + "_matrix", "api:grpc", "routes_found:true"};
-            tags = applyRegionName(tags, regionName);
-            sendDatadogStats(statsDClient, tags, durationSeconds);
-
-            MatrixRouteReply result = MatrixRouteReply.newBuilder().addAllTimes(timeRows).addAllDistances(distanceRows).build();
-            responseObserver.onNext(result);
-            responseObserver.onCompleted();
-        } catch (Exception e) {
-            logger.error("Error while completing GraphHopper matrix request! ", e);
-
-            double durationSeconds = (System.currentTimeMillis() - startTime) / 1000.0;
-            String[] tags = {"mode:" + request.getMode() + "_matrix", "api:grpc", "routes_found:false"};
-            tags = applyRegionName(tags, regionName);
-            sendDatadogStats(statsDClient, tags, durationSeconds);
-
-            Status status = Status.newBuilder()
-                    .setCode(Code.INTERNAL.getNumber())
-                    .setMessage("GH internal error! Matrix request could not be completed.")
-                    .build();
-            responseObserver.onError(StatusProto.toStatusRuntimeException(status));
-        }
+        this.streetRouter = new StreetRouter(graphHopper, statsDClient, regionName);
+        this.transitRouter = new TransitRouter(graphHopper, ptRouter, gtfsLinkMappings,
+                gtfsRouteInfo, gtfsFeedIdMapping, statsDClient, regionName);
+        this.matrixRouter = new MatrixRouter(matrixAPI, statsDClient, regionName);
+        this.isochroneRouter = new IsochroneRouter(graphHopper, new JTSTriangulator(graphHopper.getRouterConfig()));
     }
 
     @Override
@@ -289,447 +66,22 @@ public class RouterImpl extends router.RouterGrpc.RouterImplBase {
     }
 
     @Override
+    public void routeStreetMode(StreetRouteRequest request, StreamObserver<StreetRouteReply> responseObserver) {
+        streetRouter.routeStreetMode(request, responseObserver);
+    }
+
+    @Override
+    public void routeMatrix(router.RouterOuterClass.MatrixRouteRequest request, StreamObserver<router.RouterOuterClass.MatrixRouteReply> responseObserver) {
+        matrixRouter.routeMatrix(request, responseObserver);
+    }
+
+    @Override
     public void routePt(PtRouteRequest request, StreamObserver<PtRouteReply> responseObserver) {
-        long startTime = System.currentTimeMillis();
-
-        Point fromPoint = request.getPoints(0);
-        Point toPoint = request.getPoints(1);
-
-        Request ghPtRequest = new Request(fromPoint.getLat(), fromPoint.getLon(), toPoint.getLat(), toPoint.getLon());
-        ghPtRequest.setEarliestDepartureTime(Instant.ofEpochSecond(
-                request.getEarliestDepartureTime().getSeconds(), request.getEarliestDepartureTime().getNanos()
-        ));
-        ghPtRequest.setLimitSolutions(request.getLimitSolutions());
-        ghPtRequest.setLocale(Locale.US);
-        ghPtRequest.setArriveBy(false);
-        ghPtRequest.setPathDetails(Lists.newArrayList("stable_edge_ids"));
-        ghPtRequest.setProfileQuery(true);
-        ghPtRequest.setMaxProfileDuration(Duration.ofMinutes(request.getMaxProfileDuration()));
-        ghPtRequest.setBetaWalkTime(request.getBetaWalkTime());
-        ghPtRequest.setLimitStreetTime(Duration.ofSeconds(request.getLimitStreetTimeSeconds()));
-        ghPtRequest.setIgnoreTransfers(!request.getUsePareto()); // ignoreTransfers=true means pareto queries are off
-        ghPtRequest.setBetaTransfers(request.getBetaTransfers());
-
-        try {
-            GHResponse ghResponse = ptRouter.route(ghPtRequest);
-            List<ResponsePath> pathsWithStableIds = Lists.newArrayList();
-            for (ResponsePath path : ghResponse.getAll()) {
-                // Ignore walking-only responses, because we route those separately from PT
-                if (path.getLegs().size() == 1 && path.getLegs().get(0).type.equals("walk")) {
-                    continue;
-                }
-
-                // Replace the path's legs with newly-constructed legs containing stable edge IDs
-                ArrayList<Trip.Leg> legs = new ArrayList<>(path.getLegs());
-                path.getLegs().clear();
-
-                for (int i = 0; i < legs.size(); i++) {
-                    Trip.Leg leg = legs.get(i);
-                    if (leg instanceof Trip.WalkLeg) {
-                        Trip.WalkLeg thisLeg = (Trip.WalkLeg) leg;
-                        String travelSegmentType;
-                        // We only expect graphhopper to return ACCESS + EGRESS walk legs
-                        if (i == 0) {
-                            travelSegmentType = "ACCESS";
-                        } else {
-                            travelSegmentType = "EGRESS";
-                        }
-                        path.getLegs().add(new CustomWalkLeg(thisLeg, fetchWalkLegStableIds(thisLeg), travelSegmentType));
-                    } else if (leg instanceof Trip.PtLeg) {
-                        Trip.PtLeg thisLeg = (Trip.PtLeg) leg;
-                        path.getLegs().add(getCustomPtLeg(thisLeg));
-
-                        // If this PT leg is followed by another PT leg, add a TRANSFER walk leg between them
-                        if (i < legs.size() - 1 && legs.get(i + 1) instanceof Trip.PtLeg) {
-                            Trip.PtLeg nextLeg = (Trip.PtLeg) legs.get(i + 1);
-                            Trip.Stop lastStopOfThisLeg = thisLeg.stops.get(thisLeg.stops.size() - 1);
-                            Trip.Stop firstStopOfNextLeg = nextLeg.stops.get(0);
-                            if (!lastStopOfThisLeg.stop_id.equals(firstStopOfNextLeg.stop_id)) {
-                                GHRequest r = new GHRequest(
-                                        lastStopOfThisLeg.geometry.getY(), lastStopOfThisLeg.geometry.getX(),
-                                        firstStopOfNextLeg.geometry.getY(), firstStopOfNextLeg.geometry.getX());
-                                r.setProfile("foot");
-                                r.setPathDetails(Arrays.asList("stable_edge_ids"));
-                                GHResponse transfer = graphHopper.route(r);
-                                if (!transfer.hasErrors()) {
-                                    ResponsePath transferPath = transfer.getBest();
-                                    Trip.WalkLeg transferLeg = new Trip.WalkLeg(
-                                            lastStopOfThisLeg.stop_name,
-                                            thisLeg.getArrivalTime(),
-                                            transferPath.getPoints().getCachedLineString(false),
-                                            transferPath.getDistance(),
-                                            transferPath.getInstructions(),
-                                            transferPath.getPathDetails(),
-                                            Date.from(thisLeg.getArrivalTime().toInstant().plusMillis(transferPath.getTime()))
-                                    );
-                                    path.getLegs().add(new CustomWalkLeg(transferLeg, fetchWalkLegStableIds(transferLeg), "TRANSFER"));
-                                }
-                            }
-                        }
-                    }
-                }
-
-                // ACCESS legs contains stable IDs for both ACCESS and EGRESS legs for some reason,
-                // so we remove the EGRESS leg IDs from the ACCESS leg before storing the path
-                CustomWalkLeg accessLeg = (CustomWalkLeg) path.getLegs().get(0);
-                CustomWalkLeg egressLeg = (CustomWalkLeg) path.getLegs().get(path.getLegs().size() - 1);
-                accessLeg.stableEdgeIds.removeAll(egressLeg.stableEdgeIds);
-
-                // Calculate correct distance incorporating foot + pt legs
-                path.setDistance(path.getLegs().stream().mapToDouble(l -> l.distance).sum());
-
-                path.getPathDetails().clear();
-                pathsWithStableIds.add(path);
-            }
-
-            if (pathsWithStableIds.size() == 0) {
-                String message = "Transit path could not be found between " + fromPoint.getLat() + "," +
-                        fromPoint.getLon() + " to " + toPoint.getLat() + "," + toPoint.getLon();
-                // logger.warn(message);
-
-                double durationSeconds = (System.currentTimeMillis() - startTime) / 1000.0;
-                String[] tags = {"mode:pt", "api:grpc", "routes_found:false"};
-                tags = applyRegionName(tags, regionName);
-                sendDatadogStats(statsDClient, tags, durationSeconds);
-
-                Status status = Status.newBuilder()
-                        .setCode(Code.NOT_FOUND.getNumber())
-                        .setMessage(message)
-                        .build();
-                responseObserver.onError(StatusProto.toStatusRuntimeException(status));
-            } else {
-                PtRouteReply.Builder replyBuilder = PtRouteReply.newBuilder();
-                for (ResponsePath responsePath : pathsWithStableIds) {
-                    List<PtLeg> legs = responsePath.getLegs().stream()
-                            .map(RouterImpl::createPtLeg)
-                            .collect(toList());
-
-                    replyBuilder.addPaths(PtPath.newBuilder()
-                            .setDurationMillis(responsePath.getTime())
-                            .setDistanceMeters(responsePath.getDistance())
-                            .setTransfers(responsePath.getNumChanges())
-                            .addAllLegs(legs)
-                    );
-                }
-
-                double durationSeconds = (System.currentTimeMillis() - startTime) / 1000.0;
-                String[] tags = {"mode:pt", "api:grpc", "routes_found:true"};
-                tags = applyRegionName(tags, regionName);
-                sendDatadogStats(statsDClient, tags, durationSeconds);
-
-                responseObserver.onNext(replyBuilder.build());
-                responseObserver.onCompleted();
-            }
-        } catch (PointNotFoundException e) {
-            String message = "Path could not be found between " + fromPoint.getLat() + "," +
-                    fromPoint.getLon() + " to " + toPoint.getLat() + "," + toPoint.getLon() +
-                    "; one or both endpoints could not be snapped to a road segment";
-            // logger.warn(message);
-
-            double durationSeconds = (System.currentTimeMillis() - startTime) / 1000.0;
-            String[] tags = {"mode:pt", "api:grpc", "routes_found:false"};
-            tags = applyRegionName(tags, regionName);
-            sendDatadogStats(statsDClient, tags, durationSeconds);
-
-            Status status = Status.newBuilder()
-                    .setCode(Code.NOT_FOUND.getNumber())
-                    .setMessage(message)
-                    .build();
-            responseObserver.onError(StatusProto.toStatusRuntimeException(status));
-        } catch (Exception e) {
-            logger.error("GraphHopper internal error! ", e);
-
-            double durationSeconds = (System.currentTimeMillis() - startTime) / 1000.0;
-            String[] tags = {"mode:pt", "api:grpc", "routes_found:error"};
-            tags = applyRegionName(tags, regionName);
-            sendDatadogStats(statsDClient, tags, durationSeconds);
-
-            Status status = Status.newBuilder()
-                    .setCode(Code.INTERNAL.getNumber())
-                    .setMessage("GH internal error! Path could not be found between " + fromPoint.getLat() + "," +
-                            fromPoint.getLon() + " to " + toPoint.getLat() + "," + toPoint.getLon())
-                    .build();
-            responseObserver.onError(StatusProto.toStatusRuntimeException(status));
-        }
-    }
-
-    private static List<String> fetchWalkLegStableIds(Trip.WalkLeg leg) {
-        return leg.details.get("stable_edge_ids").stream()
-                .map(idPathDetail -> (String) idPathDetail.getValue())
-                .filter(id -> id.length() == 20)
-                .collect(toList());
-    }
-
-    private static PtLeg createPtLeg(Trip.Leg leg) {
-        if (leg.type.equals("walk")) {
-            CustomWalkLeg walkLeg = (CustomWalkLeg) leg;
-            return PtLeg.newBuilder()
-                    .setDepartureTime(Timestamp.newBuilder()
-                            .setSeconds(walkLeg.getDepartureTime().getTime() / 1000) // getTime() returns millis
-                            .build())
-                    .setArrivalTime(Timestamp.newBuilder()
-                            .setSeconds(walkLeg.getArrivalTime().getTime() / 1000) // getTime() returns millis
-                            .build())
-                    .setDistanceMeters(walkLeg.getDistance())
-                    .addAllStableEdgeIds(walkLeg.stableEdgeIds)
-                    .setTravelSegmentType(walkLeg.travelSegmentType)
-                    .build();
-        } else { // leg is a PT leg
-            CustomPtLeg ptLeg = (CustomPtLeg) leg;
-            TransitMetadata ptMetadata = TransitMetadata.newBuilder()
-                    .setTripId(ptLeg.trip_id)
-                    .setRouteId(ptLeg.route_id)
-                    .setAgencyName(ptLeg.agencyName)
-                    .setRouteShortName(ptLeg.routeShortName)
-                    .setRouteLongName(ptLeg.routeLongName)
-                    .setRouteType(ptLeg.routeType)
-                    .setDirection(ptLeg.trip_headsign)
-                    .addAllStops(ptLeg.stops.stream().map(stop -> Stop.newBuilder()
-                            .setStopId(stop.stop_id)
-                            .setStopName(stop.stop_name)
-                            .setArrivalTime(stop.arrivalTime == null ? Timestamp.newBuilder().build()
-                                    : Timestamp.newBuilder().setSeconds(stop.arrivalTime.getTime() / 1000).build())
-                            .setDepartureTime(stop.departureTime == null ? Timestamp.newBuilder().build()
-                                    : Timestamp.newBuilder().setSeconds(stop.departureTime.getTime() / 1000).build())
-                            .setPoint(Point.newBuilder().setLat(stop.geometry.getY()).setLon(stop.geometry.getX()).build())
-                            .build()).collect(toList())
-                    ).build();
-            return PtLeg.newBuilder()
-                    .setDepartureTime(Timestamp.newBuilder()
-                            .setSeconds(ptLeg.getDepartureTime().getTime() / 1000) // getTime() returns millis
-                            .build())
-                    .setArrivalTime(Timestamp.newBuilder()
-                            .setSeconds(ptLeg.getArrivalTime().getTime() / 1000) // getTime() returns millis
-                            .build())
-                    .setDistanceMeters(ptLeg.getDistance())
-                    .addAllStableEdgeIds(ptLeg.stableEdgeIds)
-                    .setTransitMetadata(ptMetadata)
-                    .build();
-        }
-    }
-
-    public static class CustomWalkLeg extends Trip.WalkLeg {
-        public final List<String> stableEdgeIds;
-        public final String type;
-        public final String travelSegmentType;
-
-        public CustomWalkLeg(Trip.WalkLeg leg, List<String> stableEdgeIds, String travelSegmentType) {
-            super(leg.departureLocation, leg.getDepartureTime(), leg.geometry,
-                    leg.distance, leg.instructions, leg.details, leg.getArrivalTime());
-            this.stableEdgeIds = stableEdgeIds;
-            this.details.clear();
-            this.type = "foot";
-            this.travelSegmentType = travelSegmentType;
-        }
-    }
-
-    // Create new version of PtLeg class that stores stable edge IDs in class var;
-    // this var will automatically get added to JSON response
-    public static class CustomPtLeg extends Trip.PtLeg {
-        public final List<String> stableEdgeIds;
-        public final String agencyName;
-        public final String routeShortName;
-        public final String routeLongName;
-        public final String routeType;
-
-        public CustomPtLeg(Trip.PtLeg leg, List<String> stableEdgeIds, List<Trip.Stop> updatedStops, double distance,
-                           String agencyName, String routeShortName, String routeLongName, String routeType) {
-            super(leg.feed_id, leg.isInSameVehicleAsPrevious, leg.trip_id, leg.route_id,
-                    leg.trip_headsign, updatedStops, distance, leg.travelTime, leg.geometry);
-            this.stableEdgeIds = stableEdgeIds;
-            this.agencyName = agencyName;
-            this.routeShortName = routeShortName;
-            this.routeLongName = routeLongName;
-            this.routeType = routeType;
-        }
-    }
-
-    private CustomPtLeg getCustomPtLeg(Trip.PtLeg leg) {
-        // Ordered list of GTFS route info, containing agency_name, route_short_name, route_long_name, route_type
-        List<String> routeInfo = gtfsRouteInfo.getOrDefault(gtfsRouteInfoKey(leg), Lists.newArrayList("", "", "", ""));
-        String routeType = routeInfo.get(3);
-
-        List<Trip.Stop> stops = leg.stops;
-        double legDistance = 0.0;
-        List<String> stableEdgeIdSegments = Lists.newArrayList();
-        for (int i = 0; i < stops.size() - 1; i++) {
-            Trip.Stop from = stops.get(i);
-            Trip.Stop to = stops.get(i + 1);
-            legDistance += DistanceCalcEarth.DIST_EARTH.calcDist(
-                    from.geometry.getY(), from.geometry.getX(), to.geometry.getY(), to.geometry.getX()
-            );
-
-            if (STREET_BASED_ROUTE_TYPES.contains(Integer.parseInt(routeType))) {
-                // Retrieve stable edge IDs for each stop->stop segment of leg
-                String stopPair = gtfsFeedIdMapping.get(leg.feed_id) + ":" + from.stop_id + "," + to.stop_id;
-                if (gtfsLinkMappings.containsKey(stopPair)) {
-                    if (!gtfsLinkMappings.get(stopPair).isEmpty()) {
-                        stableEdgeIdSegments.add(gtfsLinkMappings.get(stopPair));
-                    }
-                }
-            }
-        }
-         List<String> stableEdgeIdsList = stableEdgeIdSegments.stream()
-                    .flatMap(segment -> Arrays.stream(segment.split(",")))
-                    .collect(toList());
-
-        // Remove duplicates from stable ID list while retaining order;
-        // needed because start/end of sequential segments overlap by 1 edge
-        Set<String> stableEdgeIdsWithoutDuplicates = Sets.newLinkedHashSet(stableEdgeIdsList);
-        stableEdgeIdsList.clear();
-        stableEdgeIdsList.addAll(stableEdgeIdsWithoutDuplicates);
-
-        // Convert any missing info to empty string to prevent NPE
-        routeInfo = routeInfo.stream().map(info -> info == null ? "" : info).collect(toList());
-
-        if (!gtfsRouteInfo.containsKey(gtfsRouteInfoKey(leg))) {
-            logger.info("Failed to find route info for route " + leg.route_id + " for PT trip leg " + leg.toString());
-        }
-
-        // Add proper GTFS feed ID as prefix to all stop names in Leg
-        List<Trip.Stop> updatedStops = Lists.newArrayList();
-        for (Trip.Stop stop : leg.stops) {
-            String updatedStopId = gtfsFeedIdMapping.get(leg.feed_id) + ":" + stop.stop_id;
-            updatedStops.add(new Trip.Stop(updatedStopId, stop.stop_name, stop.geometry, stop.arrivalTime,
-                    stop.plannedArrivalTime, stop.predictedArrivalTime, stop.arrivalCancelled, stop.departureTime,
-                    stop.plannedDepartureTime, stop.predictedDepartureTime, stop.departureCancelled));
-        }
-
-        return new CustomPtLeg(leg, stableEdgeIdsList, updatedStops, legDistance,
-                routeInfo.get(0), routeInfo.get(1), routeInfo.get(2), routeType);
+        transitRouter.routePt(request, responseObserver);
     }
 
     @Override
     public void routeIsochrone(IsochroneRouteRequest request, StreamObserver<IsochroneRouteReply> responseObserver) {
-        PMap hintsMap = new PMap();
-        hintsMap.putObject(Parameters.CH.DISABLE, true);
-
-        String mode = request.getMode();
-        Profile profile = graphHopper.getProfile(mode);
-        if (profile == null) {
-            throw new IllegalArgumentException("The requested mode '" + mode + "' does not exist");
-        }
-        FlagEncoder encoder = graphHopper.getEncodingManager().getEncoder(profile.getVehicle());
-        EdgeFilter edgeFilter = DefaultEdgeFilter.allEdges(encoder);
-        LocationIndex locationIndex = graphHopper.getLocationIndex();
-        Snap snap = locationIndex.findClosest(request.getCenter().getLat(), request.getCenter().getLon(), edgeFilter);
-        if (!snap.isValid())
-            throw new IllegalArgumentException("Point not found:" + request.getCenter());
-
-        Graph graph = graphHopper.getGraphHopperStorage();
-        QueryGraph queryGraph = QueryGraph.create(graph, snap);
-        Weighting weighting = graphHopper.createWeighting(profile, hintsMap);
-        TraversalMode traversalMode = profile.isTurnCosts() ? EDGE_BASED : NODE_BASED;
-        ShortestPathTree shortestPathTree = new ShortestPathTree(queryGraph, weighting, request.getReverseFlow(), traversalMode);
-
-        double limit;
-        if (request.getWeightLimit() > 0) {
-            limit = request.getWeightLimit();
-            shortestPathTree.setWeightLimit(limit + Math.max(limit * 0.14, 2_000));
-        } else if (request.getDistanceLimit() > 0) {
-            limit = request.getDistanceLimit();
-            shortestPathTree.setDistanceLimit(limit + Math.max(limit * 0.14, 2_000));
-        } else {
-            limit = request.getTimeLimit() * 1000;
-            shortestPathTree.setTimeLimit(limit + Math.max(limit * 0.14, 200_000));
-        }
-        ArrayList<Double> zs = new ArrayList<>();
-        double delta = limit / request.getNBuckets();
-        for (int i = 0; i < request.getNBuckets(); i++) {
-            zs.add((i + 1) * delta);
-        }
-
-        ToDoubleFunction<ShortestPathTree.IsoLabel> fz;
-        if (request.getWeightLimit() > 0) {
-            fz = l -> l.weight;
-        } else if (request.getDistanceLimit() > 0) {
-            fz = l -> l.distance;
-        } else {
-            fz = l -> l.time;
-        }
-
-        Triangulator.Result result = isochroneTriangulator.triangulate(snap, queryGraph, shortestPathTree, fz, degreesFromMeters(request.getTolerance()));
-
-        ContourBuilder contourBuilder = new ContourBuilder(result.triangulation);
-        ArrayList<Geometry> isochrones = new ArrayList<>();
-        for (Double z : zs) {
-            logger.info("Building contour z={}", z);
-            MultiPolygon isochrone = contourBuilder.computeIsoline(z, result.seedEdges);
-            if (!isochrone.isEmpty()) {
-                if (request.getFullGeometry()) {
-                    isochrones.add(isochrone);
-                } else {
-                    Polygon maxPolygon = heuristicallyFindMainConnectedComponent(
-                            isochrone,
-                            isochrone.getFactory().createPoint(
-                                    new Coordinate(request.getCenter().getLon(), request.getCenter().getLat())
-                            )
-                    );
-                    isochrones.add(isochrone.getFactory().createPolygon(((LinearRing) maxPolygon.getExteriorRing())));
-                }
-            }
-        }
-        IsochroneRouteReply.Builder replyBuilder = IsochroneRouteReply.newBuilder();
-        for (int i = 0; i < isochrones.size(); i++) {
-            Geometry isochrone = isochrones.get(i);
-            replyBuilder.addBuckets(IsochroneBucket.newBuilder()
-                    .setBucket(i)
-                    .setGeometry(isochrone.toString())
-            );
-        }
-        responseObserver.onNext(replyBuilder.build());
-        responseObserver.onCompleted();
-    }
-
-    // Copied from IsochroneResource.java
-    private Polygon heuristicallyFindMainConnectedComponent(MultiPolygon multiPolygon, org.locationtech.jts.geom.Point point) {
-        int maxPoints = 0;
-        Polygon maxPolygon = null;
-        for (int j = 0; j < multiPolygon.getNumGeometries(); j++) {
-            Polygon polygon = (Polygon) multiPolygon.getGeometryN(j);
-            if (polygon.contains(point)) {
-                return polygon;
-            }
-            if (polygon.getNumPoints() > maxPoints) {
-                maxPoints = polygon.getNumPoints();
-                maxPolygon = polygon;
-            }
-        }
-        return maxPolygon;
-    }
-
-    /**
-     * We want to specify a tolerance in something like meters, but we need it in unprojected lat/lon-space.
-     * This is more correct in some parts of the world, and in some directions, than in others.
-     *
-     * Copied from IsochroneResource.java
-     *
-     * @param distanceInMeters distance in meters
-     * @return "distance" in degrees
-     */
-    static double degreesFromMeters(double distanceInMeters) {
-        return distanceInMeters / DistanceCalcEarth.METERS_PER_DEGREE;
-    }
-
-    private static String gtfsRouteInfoKey(Trip.PtLeg leg) {
-        return leg.feed_id + ":" + leg.route_id;
-    }
-
-    private static void sendDatadogStats(StatsDClient statsDClient, String[] tags, double durationSeconds) {
-        if (statsDClient != null) {
-            statsDClient.incrementCounter("routers.num_requests", tags);
-            statsDClient.distribution("routers.request_seconds", durationSeconds, tags);
-        }
-    }
-
-    // If a region name has been set, add it to tag list
-    private static String[] applyRegionName(String[] tags, String regionName) {
-        if (regionName == null) {
-            return tags;
-        } else {
-            List<String> newTags = Lists.newArrayList(tags);
-            newTags.add("replica_region:" + regionName);
-            return newTags.toArray(new String[newTags.size()]);
-        }
+        isochroneRouter.routeIsochrone(request, responseObserver);
     }
 }
