+++ grpc/src/main/java/com/replica/api/TransitRouter.java
@@ -1,0 +1,369 @@
+package com.replica.api;
+
+import com.google.common.collect.Lists;
+import com.google.common.collect.Sets;
+import com.google.protobuf.Timestamp;
+import com.google.rpc.Code;
+import com.google.rpc.Status;
+import com.graphhopper.*;
+import com.graphhopper.gtfs.PtRouter;
+import com.graphhopper.gtfs.Request;
+import com.graphhopper.util.DistanceCalcEarth;
+import com.graphhopper.util.exceptions.PointNotFoundException;
+import com.timgroup.statsd.StatsDClient;
+import io.grpc.protobuf.StatusProto;
+import io.grpc.stub.StreamObserver;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import router.RouterOuterClass;
+
+import java.time.Duration;
+import java.time.Instant;
+import java.util.*;
+
+import static com.replica.metrics.MetricsHelper.applyRegionName;
+import static com.replica.metrics.MetricsHelper.sendDatadogStats;
+import static java.util.stream.Collectors.toList;
+
+public class TransitRouter {
+
+    final Set<Integer> STREET_BASED_ROUTE_TYPES = Sets.newHashSet(0, 3, 5);
+
+    private static final Logger logger = LoggerFactory.getLogger(TransitRouter.class);
+    private final GraphHopper graphHopper;
+    private final PtRouter ptRouter;
+    private Map<String, String> gtfsLinkMappings;
+    private Map<String, List<String>> gtfsRouteInfo;
+    private Map<String, String> gtfsFeedIdMapping;
+    private final StatsDClient statsDClient;
+    private String regionName;
+
+    public TransitRouter(GraphHopper graphHopper, PtRouter ptRouter,
+                         Map<String, String> gtfsLinkMappings,
+                         Map<String, List<String>> gtfsRouteInfo,
+                         Map<String, String> gtfsFeedIdMapping,
+                         StatsDClient statsDClient,
+                         String regionName) {
+        this.graphHopper = graphHopper;
+        this.ptRouter = ptRouter;
+        this.gtfsLinkMappings = gtfsLinkMappings;
+        this.gtfsRouteInfo = gtfsRouteInfo;
+        this.gtfsFeedIdMapping = gtfsFeedIdMapping;
+        this.statsDClient = statsDClient;
+        this.regionName = regionName;
+    }
+
+    public void routePt(RouterOuterClass.PtRouteRequest request, StreamObserver<RouterOuterClass.PtRouteReply> responseObserver) {
+        long startTime = System.currentTimeMillis();
+
+        RouterOuterClass.Point fromPoint = request.getPoints(0);
+        RouterOuterClass.Point toPoint = request.getPoints(1);
+
+        Request ghPtRequest = new Request(fromPoint.getLat(), fromPoint.getLon(), toPoint.getLat(), toPoint.getLon());
+        ghPtRequest.setEarliestDepartureTime(Instant.ofEpochSecond(
+                request.getEarliestDepartureTime().getSeconds(), request.getEarliestDepartureTime().getNanos()
+        ));
+        ghPtRequest.setLimitSolutions(request.getLimitSolutions());
+        ghPtRequest.setLocale(Locale.US);
+        ghPtRequest.setArriveBy(false);
+        ghPtRequest.setPathDetails(Lists.newArrayList("stable_edge_ids"));
+        ghPtRequest.setProfileQuery(true);
+        ghPtRequest.setMaxProfileDuration(Duration.ofMinutes(request.getMaxProfileDuration()));
+        ghPtRequest.setBetaWalkTime(request.getBetaWalkTime());
+        ghPtRequest.setLimitStreetTime(Duration.ofSeconds(request.getLimitStreetTimeSeconds()));
+        ghPtRequest.setIgnoreTransfers(!request.getUsePareto()); // ignoreTransfers=true means pareto queries are off
+        ghPtRequest.setBetaTransfers(request.getBetaTransfers());
+
+        try {
+            GHResponse ghResponse = ptRouter.route(ghPtRequest);
+            List<ResponsePath> pathsWithStableIds = Lists.newArrayList();
+            for (ResponsePath path : ghResponse.getAll()) {
+                // Ignore walking-only responses, because we route those separately from PT
+                if (path.getLegs().size() == 1 && path.getLegs().get(0).type.equals("walk")) {
+                    continue;
+                }
+
+                // Replace the path's legs with newly-constructed legs containing stable edge IDs
+                ArrayList<Trip.Leg> legs = new ArrayList<>(path.getLegs());
+                path.getLegs().clear();
+
+                for (int i = 0; i < legs.size(); i++) {
+                    Trip.Leg leg = legs.get(i);
+                    if (leg instanceof Trip.WalkLeg) {
+                        Trip.WalkLeg thisLeg = (Trip.WalkLeg) leg;
+                        String travelSegmentType;
+                        // We only expect graphhopper to return ACCESS + EGRESS walk legs
+                        if (i == 0) {
+                            travelSegmentType = "ACCESS";
+                        } else {
+                            travelSegmentType = "EGRESS";
+                        }
+                        path.getLegs().add(new CustomWalkLeg(thisLeg, fetchWalkLegStableIds(thisLeg), travelSegmentType));
+                    } else if (leg instanceof Trip.PtLeg) {
+                        Trip.PtLeg thisLeg = (Trip.PtLeg) leg;
+                        path.getLegs().add(getCustomPtLeg(thisLeg));
+
+                        // If this PT leg is followed by another PT leg, add a TRANSFER walk leg between them
+                        if (i < legs.size() - 1 && legs.get(i + 1) instanceof Trip.PtLeg) {
+                            Trip.PtLeg nextLeg = (Trip.PtLeg) legs.get(i + 1);
+                            Trip.Stop lastStopOfThisLeg = thisLeg.stops.get(thisLeg.stops.size() - 1);
+                            Trip.Stop firstStopOfNextLeg = nextLeg.stops.get(0);
+                            if (!lastStopOfThisLeg.stop_id.equals(firstStopOfNextLeg.stop_id)) {
+                                GHRequest r = new GHRequest(
+                                        lastStopOfThisLeg.geometry.getY(), lastStopOfThisLeg.geometry.getX(),
+                                        firstStopOfNextLeg.geometry.getY(), firstStopOfNextLeg.geometry.getX());
+                                r.setProfile("foot");
+                                r.setPathDetails(Arrays.asList("stable_edge_ids"));
+                                GHResponse transfer = graphHopper.route(r);
+                                if (!transfer.hasErrors()) {
+                                    ResponsePath transferPath = transfer.getBest();
+                                    Trip.WalkLeg transferLeg = new Trip.WalkLeg(
+                                            lastStopOfThisLeg.stop_name,
+                                            thisLeg.getArrivalTime(),
+                                            transferPath.getPoints().getCachedLineString(false),
+                                            transferPath.getDistance(),
+                                            transferPath.getInstructions(),
+                                            transferPath.getPathDetails(),
+                                            Date.from(thisLeg.getArrivalTime().toInstant().plusMillis(transferPath.getTime()))
+                                    );
+                                    path.getLegs().add(new CustomWalkLeg(transferLeg, fetchWalkLegStableIds(transferLeg), "TRANSFER"));
+                                }
+                            }
+                        }
+                    }
+                }
+
+                // ACCESS legs contains stable IDs for both ACCESS and EGRESS legs for some reason,
+                // so we remove the EGRESS leg IDs from the ACCESS leg before storing the path
+                CustomWalkLeg accessLeg = (CustomWalkLeg) path.getLegs().get(0);
+                CustomWalkLeg egressLeg = (CustomWalkLeg) path.getLegs().get(path.getLegs().size() - 1);
+                accessLeg.stableEdgeIds.removeAll(egressLeg.stableEdgeIds);
+
+                // Calculate correct distance incorporating foot + pt legs
+                path.setDistance(path.getLegs().stream().mapToDouble(l -> l.distance).sum());
+
+                path.getPathDetails().clear();
+                pathsWithStableIds.add(path);
+            }
+
+            if (pathsWithStableIds.size() == 0) {
+                String message = "Transit path could not be found between " + fromPoint.getLat() + "," +
+                        fromPoint.getLon() + " to " + toPoint.getLat() + "," + toPoint.getLon();
+                // logger.warn(message);
+
+                double durationSeconds = (System.currentTimeMillis() - startTime) / 1000.0;
+                String[] tags = {"mode:pt", "api:grpc", "routes_found:false"};
+                tags = applyRegionName(tags, regionName);
+                sendDatadogStats(statsDClient, tags, durationSeconds);
+
+                Status status = Status.newBuilder()
+                        .setCode(Code.NOT_FOUND.getNumber())
+                        .setMessage(message)
+                        .build();
+                responseObserver.onError(StatusProto.toStatusRuntimeException(status));
+            } else {
+                RouterOuterClass.PtRouteReply.Builder replyBuilder = RouterOuterClass.PtRouteReply.newBuilder();
+                for (ResponsePath responsePath : pathsWithStableIds) {
+                    List<RouterOuterClass.PtLeg> legs = responsePath.getLegs().stream()
+                            .map(TransitRouter::createPtLeg)
+                            .collect(toList());
+
+                    replyBuilder.addPaths(RouterOuterClass.PtPath.newBuilder()
+                            .setDurationMillis(responsePath.getTime())
+                            .setDistanceMeters(responsePath.getDistance())
+                            .setTransfers(responsePath.getNumChanges())
+                            .addAllLegs(legs)
+                    );
+                }
+
+                double durationSeconds = (System.currentTimeMillis() - startTime) / 1000.0;
+                String[] tags = {"mode:pt", "api:grpc", "routes_found:true"};
+                tags = applyRegionName(tags, regionName);
+                sendDatadogStats(statsDClient, tags, durationSeconds);
+
+                responseObserver.onNext(replyBuilder.build());
+                responseObserver.onCompleted();
+            }
+        } catch (PointNotFoundException e) {
+            String message = "Path could not be found between " + fromPoint.getLat() + "," +
+                    fromPoint.getLon() + " to " + toPoint.getLat() + "," + toPoint.getLon() +
+                    "; one or both endpoints could not be snapped to a road segment";
+            // logger.warn(message);
+
+            double durationSeconds = (System.currentTimeMillis() - startTime) / 1000.0;
+            String[] tags = {"mode:pt", "api:grpc", "routes_found:false"};
+            tags = applyRegionName(tags, regionName);
+            sendDatadogStats(statsDClient, tags, durationSeconds);
+
+            Status status = Status.newBuilder()
+                    .setCode(Code.NOT_FOUND.getNumber())
+                    .setMessage(message)
+                    .build();
+            responseObserver.onError(StatusProto.toStatusRuntimeException(status));
+        } catch (Exception e) {
+            logger.error("GraphHopper internal error! ", e);
+
+            double durationSeconds = (System.currentTimeMillis() - startTime) / 1000.0;
+            String[] tags = {"mode:pt", "api:grpc", "routes_found:error"};
+            tags = applyRegionName(tags, regionName);
+            sendDatadogStats(statsDClient, tags, durationSeconds);
+
+            Status status = Status.newBuilder()
+                    .setCode(Code.INTERNAL.getNumber())
+                    .setMessage("GH internal error! Path could not be found between " + fromPoint.getLat() + "," +
+                            fromPoint.getLon() + " to " + toPoint.getLat() + "," + toPoint.getLon())
+                    .build();
+            responseObserver.onError(StatusProto.toStatusRuntimeException(status));
+        }
+    }
+
+    private static List<String> fetchWalkLegStableIds(Trip.WalkLeg leg) {
+        return leg.details.get("stable_edge_ids").stream()
+                .map(idPathDetail -> (String) idPathDetail.getValue())
+                .filter(id -> id.length() == 20)
+                .collect(toList());
+    }
+
+    private static RouterOuterClass.PtLeg createPtLeg(Trip.Leg leg) {
+        if (leg.type.equals("walk")) {
+            CustomWalkLeg walkLeg = (CustomWalkLeg) leg;
+            return RouterOuterClass.PtLeg.newBuilder()
+                    .setDepartureTime(Timestamp.newBuilder()
+                            .setSeconds(walkLeg.getDepartureTime().getTime() / 1000) // getTime() returns millis
+                            .build())
+                    .setArrivalTime(Timestamp.newBuilder()
+                            .setSeconds(walkLeg.getArrivalTime().getTime() / 1000) // getTime() returns millis
+                            .build())
+                    .setDistanceMeters(walkLeg.getDistance())
+                    .addAllStableEdgeIds(walkLeg.stableEdgeIds)
+                    .setTravelSegmentType(walkLeg.travelSegmentType)
+                    .build();
+        } else { // leg is a PT leg
+            CustomPtLeg ptLeg = (CustomPtLeg) leg;
+            RouterOuterClass.TransitMetadata ptMetadata = RouterOuterClass.TransitMetadata.newBuilder()
+                    .setTripId(ptLeg.trip_id)
+                    .setRouteId(ptLeg.route_id)
+                    .setAgencyName(ptLeg.agencyName)
+                    .setRouteShortName(ptLeg.routeShortName)
+                    .setRouteLongName(ptLeg.routeLongName)
+                    .setRouteType(ptLeg.routeType)
+                    .setDirection(ptLeg.trip_headsign)
+                    .addAllStops(ptLeg.stops.stream().map(stop -> RouterOuterClass.Stop.newBuilder()
+                            .setStopId(stop.stop_id)
+                            .setStopName(stop.stop_name)
+                            .setArrivalTime(stop.arrivalTime == null ? Timestamp.newBuilder().build()
+                                    : Timestamp.newBuilder().setSeconds(stop.arrivalTime.getTime() / 1000).build())
+                            .setDepartureTime(stop.departureTime == null ? Timestamp.newBuilder().build()
+                                    : Timestamp.newBuilder().setSeconds(stop.departureTime.getTime() / 1000).build())
+                            .setPoint(RouterOuterClass.Point.newBuilder().setLat(stop.geometry.getY()).setLon(stop.geometry.getX()).build())
+                            .build()).collect(toList())
+                    ).build();
+            return RouterOuterClass.PtLeg.newBuilder()
+                    .setDepartureTime(Timestamp.newBuilder()
+                            .setSeconds(ptLeg.getDepartureTime().getTime() / 1000) // getTime() returns millis
+                            .build())
+                    .setArrivalTime(Timestamp.newBuilder()
+                            .setSeconds(ptLeg.getArrivalTime().getTime() / 1000) // getTime() returns millis
+                            .build())
+                    .setDistanceMeters(ptLeg.getDistance())
+                    .addAllStableEdgeIds(ptLeg.stableEdgeIds)
+                    .setTransitMetadata(ptMetadata)
+                    .build();
+        }
+    }
+
+    public static class CustomWalkLeg extends Trip.WalkLeg {
+        public final List<String> stableEdgeIds;
+        public final String type;
+        public final String travelSegmentType;
+
+        public CustomWalkLeg(Trip.WalkLeg leg, List<String> stableEdgeIds, String travelSegmentType) {
+            super(leg.departureLocation, leg.getDepartureTime(), leg.geometry,
+                    leg.distance, leg.instructions, leg.details, leg.getArrivalTime());
+            this.stableEdgeIds = stableEdgeIds;
+            this.details.clear();
+            this.type = "foot";
+            this.travelSegmentType = travelSegmentType;
+        }
+    }
+
+    // Create new version of PtLeg class that stores stable edge IDs in class var;
+    // this var will automatically get added to JSON response
+    public static class CustomPtLeg extends Trip.PtLeg {
+        public final List<String> stableEdgeIds;
+        public final String agencyName;
+        public final String routeShortName;
+        public final String routeLongName;
+        public final String routeType;
+
+        public CustomPtLeg(Trip.PtLeg leg, List<String> stableEdgeIds, List<Trip.Stop> updatedStops, double distance,
+                           String agencyName, String routeShortName, String routeLongName, String routeType) {
+            super(leg.feed_id, leg.isInSameVehicleAsPrevious, leg.trip_id, leg.route_id,
+                    leg.trip_headsign, updatedStops, distance, leg.travelTime, leg.geometry);
+            this.stableEdgeIds = stableEdgeIds;
+            this.agencyName = agencyName;
+            this.routeShortName = routeShortName;
+            this.routeLongName = routeLongName;
+            this.routeType = routeType;
+        }
+    }
+
+    private CustomPtLeg getCustomPtLeg(Trip.PtLeg leg) {
+        // Ordered list of GTFS route info, containing agency_name, route_short_name, route_long_name, route_type
+        List<String> routeInfo = gtfsRouteInfo.getOrDefault(gtfsRouteInfoKey(leg), Lists.newArrayList("", "", "", ""));
+        String routeType = routeInfo.get(3);
+
+        List<Trip.Stop> stops = leg.stops;
+        double legDistance = 0.0;
+        List<String> stableEdgeIdSegments = Lists.newArrayList();
+        for (int i = 0; i < stops.size() - 1; i++) {
+            Trip.Stop from = stops.get(i);
+            Trip.Stop to = stops.get(i + 1);
+            legDistance += DistanceCalcEarth.DIST_EARTH.calcDist(
+                    from.geometry.getY(), from.geometry.getX(), to.geometry.getY(), to.geometry.getX()
+            );
+
+            if (STREET_BASED_ROUTE_TYPES.contains(Integer.parseInt(routeType))) {
+                // Retrieve stable edge IDs for each stop->stop segment of leg
+                String stopPair = gtfsFeedIdMapping.get(leg.feed_id) + ":" + from.stop_id + "," + to.stop_id;
+                if (gtfsLinkMappings.containsKey(stopPair)) {
+                    if (!gtfsLinkMappings.get(stopPair).isEmpty()) {
+                        stableEdgeIdSegments.add(gtfsLinkMappings.get(stopPair));
+                    }
+                }
+            }
+        }
+        List<String> stableEdgeIdsList = stableEdgeIdSegments.stream()
+                .flatMap(segment -> Arrays.stream(segment.split(",")))
+                .collect(toList());
+
+        // Remove duplicates from stable ID list while retaining order;
+        // needed because start/end of sequential segments overlap by 1 edge
+        Set<String> stableEdgeIdsWithoutDuplicates = Sets.newLinkedHashSet(stableEdgeIdsList);
+        stableEdgeIdsList.clear();
+        stableEdgeIdsList.addAll(stableEdgeIdsWithoutDuplicates);
+
+        // Convert any missing info to empty string to prevent NPE
+        routeInfo = routeInfo.stream().map(info -> info == null ? "" : info).collect(toList());
+
+        if (!gtfsRouteInfo.containsKey(gtfsRouteInfoKey(leg))) {
+            logger.info("Failed to find route info for route " + leg.route_id + " for PT trip leg " + leg.toString());
+        }
+
+        // Add proper GTFS feed ID as prefix to all stop names in Leg
+        List<Trip.Stop> updatedStops = Lists.newArrayList();
+        for (Trip.Stop stop : leg.stops) {
+            String updatedStopId = gtfsFeedIdMapping.get(leg.feed_id) + ":" + stop.stop_id;
+            updatedStops.add(new Trip.Stop(updatedStopId, stop.stop_name, stop.geometry, stop.arrivalTime,
+                    stop.plannedArrivalTime, stop.predictedArrivalTime, stop.arrivalCancelled, stop.departureTime,
+                    stop.plannedDepartureTime, stop.predictedDepartureTime, stop.departureCancelled));
+        }
+
+        return new CustomPtLeg(leg, stableEdgeIdsList, updatedStops, legDistance,
+                routeInfo.get(0), routeInfo.get(1), routeInfo.get(2), routeType);
+    }
+
+    private static String gtfsRouteInfoKey(Trip.PtLeg leg) {
+        return leg.feed_id + ":" + leg.route_id;
+    }
+}
